Fix triangle's isInside
TBD: Ellipse and consider line

+++ MFCproject/EllipseF.cpp
@@ -18,12 +18,13 @@ void EllipseF::Draw(CDC* dc) const
 bool EllipseF::isInside(const CPoint& P) const {
 	// save some time if the point is not even in the bounding rectangle
 	if (!Figure::isInside(P)) return false;
+
 	CPoint p1 = getP1();
 	CPoint p2 = getP2();
-	int centerX = p1.x + (p2.x - p1.x) / 2;
-	int centerY = p1.y + (p2.y - p1.y) / 2;
-	int r1 = p2.x - p1.x;
-	int r2 = p2.y - p1.y;
-	return (((P.x - centerX) * (P.x - centerX)) / (pow(max(r1, r2), 2))) +
-		(((P.y - centerY) * (P.y - centerY)) / (pow(min(r1, r2), 2))) <= 1;
+	double centerX = p1.x + (p2.x - p1.x) / 2;
+	double centerY = p1.y + (p2.y - p1.y) / 2;
+	double r1 = p2.x - p1.x;
+	double r2 = p2.y - p1.y;
+	return ((((P.x - centerX) * (P.x - centerX)) / (pow(max(r1, r2), 2))) +
+		(((P.y - centerY) * (P.y - centerY)) / (pow(min(r1, r2), 2)))) <= 1;
 }
